feat(Chanseok-Park): [BOJ, 5522] 카드 게임
https://www.acmicpc.net/problem/5522
